requirements.txt
scikit-learn
joblib
numpy
pandas
fastapi
uvicorn[standard]
streamlit
requests
pydantic
python-jose[cryptography]
python-multipart
python